moved function in util file
